SubNG/requirements.txt
fastapi==0.111.0
telethon==1.35.0
psutil==5.9.8
pycryptodome==3.20.0
uvicorn[standard]
